MF-34 moved message to notifications folder.
