+++ src/SFA.DAS.Reservations.Application.UnitTests/Reservations/Services/WhenNotifyingEmployerOfNewReservation.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 using SFA.DAS.Notifications.Api.Client;
 using SFA.DAS.Notifications.Api.Types;
 using SFA.DAS.Reservations.Application.Reservations.Services;
+using SFA.DAS.Reservations.Domain.Notifications;
 using SFA.DAS.Reservations.Domain.Reservations;
 using SFA.DAS.Testing.AutoFixture;
