+++ src/SFA.DAS.Reservations.Domain/Notifications/NotificationMessage.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace SFA.DAS.Reservations.Domain.Reservations
+namespace SFA.DAS.Reservations.Domain.Notifications
 {
     public class NotificationMessage
     {
