+++ src/SFA.DAS.Reservations.Domain/Reservations/INotificationsService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using SFA.DAS.Reservations.Domain.Notifications;
 
 namespace SFA.DAS.Reservations.Domain.Reservations
 {
